rapi/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
geographiclib==2.0
geopy==2.3.0
gmplot==1.4.1
gunicorn==20.1.0
idna==3.3
mysql-connector==2.2.9
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
